backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.12.0
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyMySQL==1.1.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.3
dj-database-url==2.3.0
pymysql==1.1.0

